extracted the game into one component to reduce redundancy
